api/requirements/base.txt
Django==1.8.15
djangorestframework==3.3.3
ipython==3.2.1
MySQL-python==1.2.5
PyYAML==3.11
django-redis-sessions==0.5.0
